refactor(vela-ext): extract shared IDB helpers, type onMessage, guard onFocusChanged
