+++ apps/vela-ext/entrypoints/utils/idb.ts
@@ -1,0 +1,14 @@
+export const DB_NAME = 'vela-offline-queue';
+export const STORE_NAME = 'vela-pending-sentences';
+export const DB_VERSION = 1;
+
+export function openDB(): Promise<IDBDatabase> {
+  return new Promise((resolve, reject) => {
+    const req = indexedDB.open(DB_NAME, DB_VERSION);
+    req.onupgradeneeded = () => {
+      req.result.createObjectStore(STORE_NAME, { autoIncrement: true, keyPath: 'id' });
+    };
+    req.onsuccess = () => resolve(req.result);
+    req.onerror = () => reject(req.error);
+  });
+}
