+++ apps/vela-ext/entrypoints/utils/pendingQueue.ts
@@ -1,17 +1,4 @@
-const DB_NAME = 'vela-offline-queue';
-const STORE_NAME = 'vela-pending-sentences';
-const DB_VERSION = 1;
-
-function openDB(): Promise<IDBDatabase> {
-  return new Promise((resolve, reject) => {
-    const req = indexedDB.open(DB_NAME, DB_VERSION);
-    req.onupgradeneeded = () => {
-      req.result.createObjectStore(STORE_NAME, { autoIncrement: true, keyPath: 'id' });
-    };
-    req.onsuccess = () => resolve(req.result);
-    req.onerror = () => reject(req.error);
-  });
-}
+import { openDB, STORE_NAME } from './idb';
 
 export async function getPendingQueueCount(): Promise<number> {
   try {
